AB4950 - Created new Umbraco.Web.BackOffice project. Moved dashboards into this new project
