+++ src/Umbraco.Web.BackOffice/Compose/BackOfficeComponent.cs
@@ -1,0 +1,23 @@
+﻿using Umbraco.Core.Composing;
+
+namespace Umbraco.Web.Runtime
+{
+    public sealed class BackOfficeComponent : IComponent
+    {
+        public BackOfficeComponent()
+        {
+
+        }
+
+        public void Initialize()
+        {
+
+        }
+
+        public void Terminate()
+        {
+
+        }
+
+    }
+}
