+++ src/Umbraco.Web.UI/Umbraco.Web.UI.csproj
@@ -128,6 +128,10 @@
       <Project>{f6de8da0-07cc-4ef2-8a59-2bc81dbb3830}</Project>
       <Name>Umbraco.PublishedCache.NuCache</Name>
     </ProjectReference>
+    <ProjectReference Include="..\Umbraco.Web.BackOffice\Umbraco.Web.BackOffice.csproj">
+      <Project>{9b95eef7-63fe-4432-8c63-166be9c1a929}</Project>
+      <Name>Umbraco.Web.BackOffice</Name>
+    </ProjectReference>
     <ProjectReference Include="..\Umbraco.Web\Umbraco.Web.csproj">
       <Project>{651e1350-91b6-44b7-bd60-7207006d7003}</Project>
       <Name>Umbraco.Web</Name>
